requirements.txt
anglerfish==2.5.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cssmin==0.2.0
dill==0.2.7.1
flake8==3.5.0
Flask==1.0
Flask-Assets==0.12
Flask-PyMongo==0.5.1
Flask-Scss==0.5
future==0.16.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10.1
jsmin==2.2.2
MarkupSafe==1.0
mccabe==0.6.1
monotonic==1.3
oauth2client==4.1.2
ply==3.8
protobuf==3.6.0
pyasn1==0.4.1
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pyflakes==1.6.0
pymongo==3.5.1
pyScss==1.3.5
python-dotenv==0.12.0
pytz==2019.3
requests==2.20.0
rsa==3.4.2
six==1.13.0
tenacity==4.5.0
urllib3==1.24.2
webassets==0.12.1
Werkzeug==0.15.3
